another commit lol

+++ MainScreen/screen.py
@@ -23,12 +23,29 @@ class MainScreen(Frame):
         self.load_elements(canvas)
 
     def load_elements(self, canvas):
+        self.txtbox_frame = Frame(canvas)
+
+        self.txt_Content = Text(self.txtbox_frame, width=52, height=35)
+        self.txt_Content.insert(END, "Click a file to preview.")
+        self.txt_Content.config(state=DISABLED)
+        self.scrbr_Content = Scrollbar(self.txtbox_frame)
+
+        self.txt_Content.config(yscrollcommand = self.scrbr_Content.set)
+        self.scrbr_Content.config(command=self.txt_Content.yview)
+
+        self.txt_Content.pack(side=LEFT)
+        self.scrbr_Content.pack(side=RIGHT, fill=Y)
+
+        self.txtbox_frame.pack(padx=10, side=LEFT)
+
+        # ==========
+        
         self.listbox_frame = Frame(canvas)
 
-        self.lst_Files = Listbox(self.listbox_frame, selectmode=SINGLE, width=160, height=18)
+        self.lst_Files = Listbox(self.listbox_frame, selectmode=SINGLE, width=85, height=18)
         self.lst_Files.bind('<Double-1>', self.open_pdf)
         self.lst_Files.bind('<<ListboxSelect>>', self.show_text)
-        self.lst_Files.pack(padx=10, pady=10, side=LEFT)
+        self.lst_Files.pack(padx=5, pady=10, side=LEFT)
 
         self.scrbr_Files = Scrollbar(self.listbox_frame)
         self.scrbr_Files.pack(side=RIGHT, fill=Y)
@@ -42,7 +59,7 @@ class MainScreen(Frame):
 
         self.buttons_frame = Frame(canvas)
 
-        self.btn_Browse = Button(self.buttons_frame, text="Add", command=lambda: threading.Thread(target=self.browse_files()).start, height=3, width=35, borderwidth=5)
+        self.btn_Browse = Button(self.buttons_frame, text="Add/Upload", command=lambda: threading.Thread(target=self.browse_files()).start, height=3, width=35, borderwidth=5)
         self.btn_Browse.pack()
 
         self.btn_Clear = Button(self.buttons_frame, text="Clear", command=lambda: self.clear_files(), height=3, width=35, borderwidth=5)
@@ -58,54 +75,46 @@ class MainScreen(Frame):
 
         # ==========
 
-        self.txtbox_frame = Frame(canvas)
-
-        self.txt_Content = Text(self.txtbox_frame, width=50, height=16)
-        self.txt_Content.insert(END, "Click a file to preview.")
-        self.txt_Content.config(state=DISABLED)
-        self.scrbr_Content = Scrollbar(self.txtbox_frame)
-
-        self.txt_Content.config(yscrollcommand = self.scrbr_Content.set)
-        self.scrbr_Content.config(command=self.txt_Content.yview)
-
-        self.txt_Content.pack(side=LEFT)
-        self.scrbr_Content.pack(side=RIGHT, fill=Y)
-
-        self.txtbox_frame.pack(padx=10, side=LEFT)
-
-        # ==========
-
         self.search_include_frame = Frame(canvas)
 
-        self.lbl_Include = Label(self.search_include_frame, text="Include the following keywords:", anchor=W)
-        self.lbl_Include.pack(pady=10)
+        self.lbl_Include = Label(self.search_include_frame, text="Search Keywords:", anchor=W)
+        self.lbl_Include.pack(anchor=W)
 
         self.txt_Include = Entry(self.search_include_frame, width=40)
-        self.txt_Include.pack()
+        self.txt_Include.pack(pady=10)
 
-        self.search_include_frame.pack(pady=10, side=TOP)
+        self.lbl_Include_bool = Label(self.search_include_frame, text="Booleans: ", anchor=W)
+        self.lbl_Include_bool.pack(anchor=W)
 
-        # self.search_include_bool_frame = Frame(canvas)
+        self.bool_Include = IntVar()
+        self.rdbtn_Include_AND = Radiobutton(self.search_include_frame, text="AND", variable=self.bool_Include, value=0)
+        self.rdbtn_Include_OR = Radiobutton(self.search_include_frame, text="OR", variable=self.bool_Include, value=1)
+        self.rdbtn_Include_AND.pack(anchor=W)
+        self.rdbtn_Include_OR.pack(anchor=W)
 
-        # self.bool_Include = IntVar()
-        # self.rdbtn_Include_AND = Radiobutton(self.search_include_bool_frame, text="AND", variable=self.bool_Include, value=0)
-        # self.rdbtn_Include_OR = Radiobutton(self.search_include_bool_frame, text="OR", variable=self.bool_Include, value=1)
-        # self.rdbtn_Include_AND.pack(side=LEFT)
-        # self.rdbtn_Include_OR.pack(side=LEFT)
+        self.lbl_Note = Label(self.search_include_frame, text="Multiple keywords must be separated \nwith commas.", anchor=W)
+        self.lbl_Note.pack(side=BOTTOM, pady=15)
 
-        # self.search_include_bool_frame.pack(side=TOP)
+        self.lbl_Count = Label(self.search_include_frame, text=f"{len(self.included_files)}/{len(self.PDF_files)} shown.", anchor=W)
+        self.lbl_Count.pack(side=BOTTOM, pady=15, anchor=W)
+
+        self.search_include_frame.pack(pady=15, side=TOP)
+
 
         # ==========
 
-        self.search_exclude_frame = Frame(canvas)
+        self.prg_Progress = ttk.Progressbar(canvas, orient=HORIZONTAL, length=200, mode="determinate", )
+        self.lbl_Progress = Label(text="Loading Test...")
 
-        self.lbl_Exclude = Label(self.search_exclude_frame, text="Exclude the following keywords:", anchor=W)
-        self.lbl_Exclude.pack(pady=10)
+        # self.search_exclude_frame = Frame(canvas)
 
-        self.txt_Exclude = Entry(self.search_exclude_frame, width=40)
-        self.txt_Exclude.pack()
+        # self.lbl_Exclude = Label(self.search_exclude_frame, text="Exclude the following keywords:", anchor=W)
+        # self.lbl_Exclude.pack(pady=10)
 
-        self.search_exclude_frame.pack(pady=10, side=TOP)
+        # self.txt_Exclude = Entry(self.search_exclude_frame, width=40)
+        # self.txt_Exclude.pack()
+
+        # self.search_exclude_frame.pack(pady=10, side=TOP)
 
         # self.search_exclude_bool_frame = Frame(canvas)
 
@@ -118,14 +127,6 @@ class MainScreen(Frame):
         # self.search_exclude_bool_frame.pack(side=TOP)
 
         # ==========
-
-        self.lbl_Note = Label(canvas, text="Multiple keywords must be separated \nwith commas.", anchor=W)
-        self.lbl_Note.pack(side=BOTTOM, pady=2)
-
-        # ===========
-
-        self.prg_Progress = ttk.Progressbar(canvas, orient=HORIZONTAL, length=200, mode="determinate", )
-        self.lbl_Progress = Label(text="Loading Test...")
 
     def browse_files(self):
         file_paths = filedialog.askopenfilenames(filetypes=[("PDF Files", "*.pdf")])
@@ -140,11 +141,16 @@ class MainScreen(Frame):
         for file in file_paths:
             i += 1
             file_name = os.path.basename(file)
-            self.update_progress(file_name, (i/j)*100)
-            if self.duplicate(file):
-                messagebox.showinfo(title="Duplicate Found", message=f"{file_name} is a duplicate. It will not be added on the list.")
-                continue
-            self.PDF_files[file] = PDF(file, file_name)
+            name_only, ext = os.path.splitext(file_name)
+            self.update_progress(f"Loading {file_name}", (i/j)*100)
+            if self.duplicate(name_only):
+                res = messagebox.askquestion(title="Duplicate Found", message=f"{file_name} is a duplicate. Do you wish to add it on the list?", icon='warning')
+                if res == 'no':
+                    continue
+                self.PDF_files[name_only].duplicate += 1
+                name_only = f"{name_only} ({(self.PDF_files[name_only].duplicate)})"
+            self.PDF_files[name_only] = PDF(file, file_name)
+        self.included_files = self.PDF_files
         self.hide_progress()
         self.update_list(self.PDF_files)
         self.revert()
@@ -153,6 +159,7 @@ class MainScreen(Frame):
         self.lst_Files.delete(0, END)
         for file in files:
             self.lst_Files.insert(END, file)
+        self.lbl_Count["text"] = f"{len(self.included_files)}/{len(self.PDF_files)} shown."
 
     def duplicate(self, file):
         if file in self.lst_Files.get(0, END):
@@ -160,57 +167,93 @@ class MainScreen(Frame):
         return False
 
     def clear_files(self):
+        if len(self.PDF_files) == 0:
+            return
+        
+        res = messagebox.askquestion(title="Clear", message=f"{len(self.PDF_files)} file/s will be removed on the list. Do you wish to continue? \n(You can reset the list after search instead of uploading the files again.)", icon='warning')
+        if res == 'no':
+            return
+        
         self.change_text("Click a file to preview.")
         self.lst_Files.delete(0, END)
         self.PDF_files = {}
         self.included_files = {}
         self.txt_Include.delete(0, END)
-        self.txt_Exclude.delete(0, END)
+        self.lbl_Count["text"] = f"{len(self.included_files)}/{len(self.PDF_files)} shown."
+
+        # self.txt_Exclude.delete(0, END)
         self.hide_revert_button()
 
     def search(self):
         file_paths = self.lst_Files.get(0, END)
         if len(file_paths) == 0:
-            messagebox.showwarning(title="Empty Search Files", message="There are no files to search.")
+            messagebox.showwarning(title="Empty Search Files", message="There are no files to search. Please upload a file or click the reset button after search.")
             return
         
-        included = StringManipulate.split_keywords(self.txt_Include.get())
-        excluded = StringManipulate.split_keywords(self.txt_Exclude.get())
+        keywords = StringManipulate.split_keywords(self.txt_Include.get())
+        # excluded = StringManipulate.split_keywords(self.txt_Exclude.get())
 
-        if len(included) == 0 and len(excluded) == 0:
+        if len(keywords) == 0:
             messagebox.showwarning(title="Empty Search Keyword", message="Please input at least one keyword to search.")
             return
 
         self.included_files = {}
 
+        i = 0
+        j = len(file_paths)
+        self.show_progress()
+
         for file in file_paths:
+            i += 1
+
             current_file = self.PDF_files[file]
-            for word in excluded:
-                if StringMatching.search(current_file.content, word):
-                    current_file.included = False
-                    break
-                current_file.included = True
 
-            if not current_file.included:
-                continue
-
-            for word in included:
-                if not StringMatching.search(current_file.content, word):
+            current_file.included = True
+            
+            if self.bool_Include.get() == 0:
+                for word in keywords:
+                    if not StringMatching.search(current_file.content, word):
+                        current_file.included = False
+                        break
+                    current_file.included = True
+            else:
+                for word in keywords:
+                    self.update_progress(f"Searching {word} in {current_file.file_name}", (i/j)*100)
+                    if StringMatching.search(current_file.content, word):
+                        current_file.included = True
+                        break
                     current_file.included = False
-                    break
-                current_file.included = True
 
             if current_file.included:
                 self.included_files[file] = current_file
 
+            # AND_indicator = True if self.bool_Include.get() == 0 else False
+
+            # for word in keywords:
+            #     current_file.included = False
+                
+            #     self.update_progress(f"Searching {word} in {current_file.file_name}", (i/j)*100)
+            #     keyword_found = StringMatching.search(current_file.content, word)
+            #     if AND_indicator and not keyword_found:
+            #         current_file.included = False
+            #         break
+            #     elif not AND_indicator and keyword_found:
+            #         current_file.included = True
+            #         break
+
+            #     current_file.included = True
+
         messagebox.showinfo(title=f"Results", message=f"Search found {len(self.included_files)} result/s.")
         self.update_list(self.included_files)
+        self.hide_progress()
         self.show_revert_button()
                 
     def open_pdf(self, event):
         get_selected = self.lst_Files.curselection()
-        path = self.lst_Files.get(get_selected)
-        os.startfile(path)
+        file_name = self.lst_Files.get(get_selected)
+        current_PDF = self.PDF_files[file_name]
+        get_path = current_PDF.file_path
+        os.startfile(get_path)
 
     def show_text(self, event):
         get_selected = self.lst_Files.curselection()
@@ -231,9 +274,9 @@ class MainScreen(Frame):
         self.prg_Progress.place(x=WINDOW_WIDTH/2, y=WINDOW_HEIGHT/2, anchor=CENTER)
         self.lbl_Progress.place(x=WINDOW_WIDTH/2, y=(WINDOW_HEIGHT/2)+30, anchor=CENTER)
 
-    def update_progress(self, file_name, percentage):
+    def update_progress(self, text, percentage):
         self.prg_Progress["value"] = percentage
-        self.lbl_Progress["text"] = f"Loading {file_name}"
+        self.lbl_Progress["text"] = text
         self.main.main_window.update_idletasks()
 
     def hide_progress(self):
@@ -247,9 +290,11 @@ class MainScreen(Frame):
         self.btn_Reset.place_forget()
         
     def revert(self):
+        self.included_files = self.PDF_files
         self.set_all_included()
         self.update_list(self.PDF_files)
         self.hide_revert_button()
+        self.lbl_Count["text"] = f"{len(self.included_files)}/{len(self.PDF_files)} shown."
         
     def set_all_included(self):
         for file in self.PDF_files:
